requirements.txt
appdirs==1.4.4
astroid==2.4.2
autopep8==1.5.3
bcrypt==3.2.0
beautifulsoup4==4.9.1
blinker==1.4
blueprint==3.4.2
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
ChatterBot==1.0.0
chatterbot-corpus==1.2.0
click==7.1.2
colorama==0.4.3
comtypes==1.1.7
DateTime==4.3
distlib==0.3.1
dnspython==2.0.0
email-validator==1.1.1
filelock==3.0.12
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.4.3
Flask-WTF==0.14.3
gunicorn==20.0.4
idna==2.10
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.16.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mathparse==0.1.2
mccabe==0.6.1
nltk==3.5
packaging==20.4
Pillow==7.2.0
Pint==0.14
PyAudio @ file:///C:/Users/lenovo/Downloads/PyAudio-0.2.11-cp38-cp38-win32.whl
pycodestyle==2.6.0
pycparser==2.20
pylint==2.5.3
pymongo==3.11.0
PyMySQL==0.9.3
pyparsing==2.4.7
pypiwin32==223
python-dateutil==2.7.5
pyttsx3==2.90
pytz==2020.1
pywin32==228
PyYAML==3.13
regex==2020.7.14
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
SpeechRecognition==3.8.1
SQLAlchemy==1.3.19
toml==0.10.1
tqdm==4.48.0
urllib3==1.25.9
virtualenv==20.0.31
Werkzeug==1.0.1
wikipedia==1.4.0
wrapt==1.12.1
WTForms==2.3.3
zope.interface==5.1.0
